moving into shared _map_scoring.py
