+++ src/konnektor/network_planners/generators/heuristic_maximal_network_generator.py
@@ -7,9 +7,8 @@ from collections.abc import Iterable
 import numpy as np
 from gufe import AtomMapper, Component, LigandNetwork
 
+from .._map_scoring import _parallel_map_scoring, _serial_map_scoring
 from ._abstract_network_generator import NetworkGenerator
-from ._parallel_mapping_pattern import _parallel_map_scoring
-from ._serial_mapping_pattern import _serial_map_scoring
 
 # Todo: is graph connectivity ensured?
